Split ticket_granting module
